requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0

# AWS
boto3==1.34.0

# Data processing and ML
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Database
sqlalchemy>=2.0.35  # Updated for Python 3.14 compatibility
pymysql==1.1.0
# psycopg2-binary==2.9.9  # Only needed for PostgreSQL, commented out for Python 3.14 compatibility

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# Security
werkzeug==3.0.1
bcrypt==4.1.2

# Utilities
python-dateutil==2.8.2
